+++ src_ts/elements/app-shell-components/sidebar-menu/app-menu.ts
