+++ src_ts/elements/app-shell-components/sidebar-menu/mixins/app-menu-mixin.ts
